test: testing infrastructure for commandline
This testing infrastructure is based on the trycmd library.

Initially I tried to base this off cfg(test), but trycmd exercises the real binary
and I could find no way to compile a different one in test mode.

So I introduced an explicit feature "fake" and then used a hack of depending on
the iroh-ctl package again in the dev dependencies to force the feature to exist.
This causes the tests to run against the fake api but the production command-line
is compiled against the real one (unless the fake feature is explicitly passed).

To make this work I had to force the resolver to be 2. You'd think that setting
the edition to 2021 in all crates would do that, but apparently it does not when
cargo workspaces are in use.

When I was running all the tests I noticed two places where the tests were unstable
due to underlying hash tables. One was my trycmd test for peer ids, which I fixed
by sorting them in the API implementation; stable output is nice anyway. The other
was in the resolver, which I fixed by sorting before compare to enforce stability.